Update sample-proposal.md
I've mostly aligned this with the proposal form.  I suggest adding learning goals, and MVP Feature set as paragraph inputs in the proposal form.  It gives the instructor a good idea about what they want to get out of the project and what they see as a Minimum Viable Product for scoping.